latest changes

+++ Gas.Domain/Entity/SalesManagement/CylinderSaleEntity.cs
@@ -11,7 +11,7 @@
         public string Shop_name { get; set; } = string.Empty;
         public string Sale_date { get; set; } = string.Empty;
         public int? Quantity { get; set; }
-        public int? Price { get; set; }
+        public decimal? Price { get; set; }
         public string Description { get; set; } = string.Empty;
         public int? Super_dealer_id { get; set; }
         public string Super_dealer_name { get; set; } = string.Empty;
@@ -25,7 +25,7 @@
         public int? Cylinder_category_id { get; set; }
         public string Cylinder_category_name { get; set; } = string.Empty;
         public int? Quantity { get; set; }
-        public int? Price { get; set; }
+        public decimal? Price { get; set; }
         public int? Cylinder_sale_id { get; set; }
     }
